// cola/week7/mission1/src/types/auth.ts
import { CommonResponse } from './common';

// 회원가입

export type RequestSignUpDto = {
  name: string;
  email: string;
  bio?: string;
  avatar?: string;
  password: string;
};

export type ResponseSignUpDto = CommonResponse<{
  id: number;
  name: string;
  email: string;
  bio: string | null;
  avatar: string | null;
  createdAt: Date;
  updatedAt: Date;
}>;

// 로그인

export type RequestSignInDto = {
  email: string;
  password: string;
};

export type ResponseSignInDto = CommonResponse<{
  id: number;
  name: string;
  accessToken: string;
  refreshToken: string;
}>;

// 내 정보 조회

export type ResponseMyInfoDto = CommonResponse<{
  id: number;
  name: string;
  email: string;
  bio: string | null;
  avatar: string | null;
  createdAt: Date;
  updatedAt: Date;
}>;

// 내 정보 수정

export type RequestEditMyInfoDto = {
  name: string;
  bio: string | null;
  avatar?: string | null;
};

export type ResponseEditMyInfotDto = ResponseMyInfoDto;